+++ backend/pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+asyncio_mode = auto
+asyncio_default_fixture_loop_scope = function
+testpaths = tests
+norecursedirs = .venv .git node_modules
+log_cli = true
+log_level = INFO
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::pytest.PytestCacheWarning
+    default
